aws_cdk_serverless_event_processor/requirements.txt
-e .
aws.cdk.aws_s3
aws.cdk.aws_dynamodb
aws.cdk.aws_lambda
aws.cdk.aws_logs
aws_cdk.aws_s3_notifications